requirements.txt
streamlit==1.33.0
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.12.2
scikit-learn==1.4.2
tensorflow==2.16.1
sqlalchemy==2.0.30
cryptography==42.0.5
PyCryptodome==3.20.0
bcrypt==4.1.2